requirements.txt
Flask==2.2.5
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.48
boto3==1.28.13
requests==2.31.0
python-decouple==3.8